chore(atomix): move raftlog builder to own class SonarLint complains that RaftLog class is too big.
